feat(users): user custom-field engine + contact profile (phase A)
Extend users into contact records, reusing the asset custom-field engine.

- Extract the asset schema validator into services/custom_fields/schema.rs
  (generic JSON-Schema subset) + add a `synced` keyword for sync-fed fields;
  services/assets/kinds.rs re-exports it, assets unchanged.
- Migration: user_field_schema (per-workspace, override-only custom-field
  definitions) + user_profiles (per user x workspace: job_title/organization/
  department + custom_fields JSONB + directory_synced). RLS, grants, audit.
- repository::user_contact: code-default schema (office_location synced,
  gender, birthday; all optional), schema get/set, profile get/upsert.
- Handlers + routes: GET/PUT /admin/user-fields (read staff, write admin,
  validates schema) and GET/PUT /users/{uuid}/profile-fields (self-or-admin,
  validates custom_fields against the effective schema, preserves directory-
  synced standard columns).
