+++ backend/src/handlers/mod.rs
@@ -71,6 +71,7 @@ pub mod ticket_merge;
 pub mod ticket_watchers;
 pub mod tickets;
 pub mod unsubscribe;
+pub mod user_contact;
 pub mod users;
 pub mod webhooks;
 pub mod workflow_states;
