+++ backend/src/services/custom_fields/mod.rs
@@ -1,0 +1,3 @@
+//! Shared custom-field engine: the JSON-Schema-subset validator for
+//! admin-defined fields, used by assets and (soon) users.
+pub mod schema;
