+++ backend/src/services/mod.rs
@@ -5,6 +5,7 @@ pub mod avatar_thumbnails;
 pub mod backup;
 pub mod channels;
 pub mod collab_ownership;
+pub mod custom_fields;
 pub mod dkim_verification;
 pub mod dns_diagnostics;
 pub mod email_queue;
